requirements.txt
streamlit
python-dotenv
groq
PyMuPDF
requests